chore: Update .gitignore to include TODO.md
The .gitignore file has been updated to include the TODO.md file. This change ensures that the TODO.md file is ignored by Git and not tracked in the repository. By adding TODO.md to the .gitignore file, we prevent it from being accidentally committed and pushed to the remote repository.