feat: Add iter_keys method to Storage trait
This commit adds the iter_keys method to the Storage trait in the storage/traits.rs file. The iter_keys method retrieves all the keys from the storage and returns them as a vector of strings. This new method enhances the functionality of the Storage trait by allowing users to iterate over all the keys in the store.